requirements.txt
streamlit>=1.35.0
protobuf>=4.25.3
google-cloud-aiplatform 
google-cloud-discoveryengine
google-adk
PyPDF2
python-dotenv 
pytest
google-genai
vertexai